add model analysis code

+++ src/evaluators/uniref_evaluator.py
@@ -13,8 +13,6 @@ import torch
 import tqdm
 
 # from src.evaluators.metrics import plot_roc_curve
-from src.data.benchmarking import generate_roc
-from src.data.utils import actmap_pipeline
 from src.evaluators import Evaluator
 from src.utils import encode_string_sequence
 from src.utils.gen_utils import generate_string_sequence
@@ -41,11 +39,12 @@ class UniRefEvaluator(Evaluator):
         max_seq_length=512,
         evalue_threshold=1,
         add_random_sequence=False,
-        nprobe=1,
+        nprobe=10,
         distance_threshold=0,
         normalize_embeddings=False,
         index_string="Flat",
         index_device="cpu",
+        output_path="",
     ):
         """
         Args:
@@ -84,7 +83,9 @@ class UniRefEvaluator(Evaluator):
         self.target_seqs: dict = target_seqs
         self.max_hmmer_hits: dict = hmmer_hits_max
 
-        self.encoding_func = encode_string_sequence if encoding_func is None else encoding_func
+        self.encoding_func = (
+            encode_string_sequence if encoding_func is None else encoding_func
+        )
         self.add_random_sequence: bool = add_random_sequence
         self.figure_path: str = figure_path
         self.nprobe: int = nprobe
@@ -100,32 +101,44 @@ class UniRefEvaluator(Evaluator):
         self.num_queries = None
         self.distance_threshold = float(distance_threshold)
         self.evalue_threshold = evalue_threshold
+        self.output_path = output_path
 
         if self.normalize_embeddings:
-            logger.info("Using comparison function >= threshold for filtration.")
+            logger.info(
+                "Using comparison function >= threshold for filtration."
+            )
             self.comp_func = np.greater_equal
         else:
-            logger.info("Using comparison function <= threshold for filtration.")
+            logger.info(
+                "Using comparison function <= threshold for filtration."
+            )
             self.comp_func = np.less_equal
 
         # root = "/xdisk/twheeler/colligan/data/prefilter/uniref_benchmark/"
         # self.max_hmmer_hits = get_hmmer_hits(f"{root}/max_hmmer_hits.txt")
 
     def filter_sequences_by_length(
-        self, names: List[str], sequences: List[str], model_class, apply_random_sequence: bool
+        self,
+        names: List[str],
+        sequences: List[str],
+        model_class,
+        apply_random_sequence: bool,
+        max_seq_length=512,
     ) -> Tuple[List[str], List[str], List[torch.Tensor]]:
         """Filters the sequences by length thresholding given the
         minimum and maximum length threshold variables"""
         embeddings = []
 
-        filtered_sequences, filtered_names = sequences.copy(), names.copy()
+        filtered_names = names.copy()
         num_removed = 0
         for name, sequence in tqdm.tqdm(zip(names, sequences)):
-            if self.max_seq_length >= len(sequence) >= self.minimum_seq_length:
+            if max_seq_length >= len(sequence) >= self.minimum_seq_length:
                 if apply_random_sequence:
                     # add 100 aminos on to the beginning
                     random_seq = generate_string_sequence(100)
-                    embed = self.compute_embedding(random_seq + sequence, model_class)
+                    embed = self.compute_embedding(
+                        random_seq + sequence, model_class
+                    )
                 else:
                     embed = self.compute_embedding(sequence, model_class)
                 # return: seq_lenxembed_dim shape
@@ -133,10 +146,12 @@ class UniRefEvaluator(Evaluator):
             else:
                 num_removed += 1
                 filtered_names.remove(name)
-                filtered_sequences.remove(sequence)
-                logger.debug(f"Removing sequence {sequence} with length {len(sequence)}")
+                # filtered_sequences.remove(sequence)
+                logger.debug(
+                    f"Removing sequence {sequence} with length {len(sequence)}"
+                )
         logger.info(f"removed {num_removed} sequences. ")
-        return filtered_names, filtered_sequences, embeddings
+        return filtered_names, embeddings
 
     def filter_hmmer_hits(self, target_names: List[str]):
         """Filters the max hmmer hits dict to include
@@ -144,16 +159,16 @@ class UniRefEvaluator(Evaluator):
         This is in order to keep our benchmarking consistent."""
 
         init_len = sum([len(x) for x in self.max_hmmer_hits.values()])
-        print(f"initial length: {init_len}")
+        print(f"Filtering HMMER hits")
         filtered_hmmer_hits = {}
 
         # TODO: make this faster
         num_missing = 0
         for queryname, targethits in self.max_hmmer_hits.items():
             hits_dict = {}
-            for targetname in targethits.keys():
+            for targetname, hits in targethits.items():
                 if targetname in target_names:
-                    hits_dict[targetname] = targethits[targetname]
+                    hits_dict[targetname] = hits
                 else:
                     num_missing += 1
 
@@ -169,7 +184,11 @@ class UniRefEvaluator(Evaluator):
 
     @torch.no_grad()
     def _calc_embeddings(
-        self, sequence_data: dict, model_class, apply_random_sequence: bool
+        self,
+        sequence_data: dict,
+        model_class,
+        apply_random_sequence: bool,
+        max_seq_length=512,
     ) -> Tuple[List[str], List[str], List[torch.Tensor]]:
         """Calculates the embeddings for the sequences by
         calling the model forward function. Filters the sequences by max/min
@@ -181,13 +200,17 @@ class UniRefEvaluator(Evaluator):
         sequences = list(sequence_data.values())
 
         logger.info("Filtering sequences by length...")
-        filtered_names, filtered_sequences, embeddings = self.filter_sequences_by_length(
-            names, sequences, model_class, apply_random_sequence
+        filtered_names, embeddings = self.filter_sequences_by_length(
+            names,
+            sequences,
+            model_class,
+            apply_random_sequence,
+            max_seq_length,
         )
 
-        assert len(filtered_names) == len(filtered_sequences) == len(embeddings)
+        assert len(filtered_names) == len(embeddings)
 
-        return filtered_names, filtered_sequences, embeddings
+        return filtered_names, embeddings
 
     def evaluate(self, model_class) -> dict:
         """Evaluation pipeline.
@@ -200,20 +223,22 @@ class UniRefEvaluator(Evaluator):
         if hasattr(model_class, "initial_seq_len"):
             self.tile_size = model_class.initial_seq_len
 
-        target_names, _, target_embeddings = self._calc_embeddings(
+        print(f"Found {(len(self.target_seqs))} targets")
+
+        target_names, target_embeddings = self._calc_embeddings(
             sequence_data=self.target_seqs,
             model_class=model_class,
             apply_random_sequence=False,
+            max_seq_length=self.max_seq_length,
         )
 
         del self.target_seqs  # remove from memory
 
-        self.filter_hmmer_hits(target_names)
-
-        query_names, _, query_embeddings = self._calc_embeddings(
+        query_names, query_embeddings = self._calc_embeddings(
             sequence_data=self.query_seqs,
             model_class=model_class,
             apply_random_sequence=self.add_random_sequence,
+            max_seq_length=self.max_seq_length,
         )
         del self.query_seqs
 
@@ -223,23 +248,14 @@ class UniRefEvaluator(Evaluator):
 
         # self.denom = len(query_names) * len(target_names)
         # self.num_queries = len(query_names)
-
-        generate_roc(
-            filename="data.txt",
-            modelhitsfile=self.output_path,
-            hmmerhits=self.max_hmmer_hits,
-            figure_path=self.figure_path,
-        )
-        # plot_roc_curve(
-        #     model_hits,
-        #     self.max_hmmer_hits,
-        #     self.normalize_embeddings,
-        #     self.distance_threshold,
-        #     self.denom,
-        #     self.figure_path,
-        #     self.comp_func,
-        #     evalue_thresholds=[1e-10],
+        # self.filter_hmmer_hits(target_names)
+        # generate_roc(
+        #     filename="data.txt",
+        #     modelhitsfile=self.output_path,
+        #     hmmerhits=self.max_hmmer_hits,
+        #     figure_path=self.figure_path,
         # )
+
         return model_hits
 
     def compute_embedding(self, sequence, model_class):
@@ -282,12 +298,15 @@ class UniRefEvaluator(Evaluator):
         logger.info("Beginning search.")
 
         t_begin = time.time()
-        self.output_path = os.path.join(os.path.dirname(self.figure_path), "similarities")
 
         if not os.path.exists(self.output_path):
             os.mkdir(self.output_path)
 
+        print(self.output_path)
+
         for i in tqdm.tqdm(range(len(queries))):
+            if os.path.exists(f"{self.output_path}/{query_names[i]}.txt"):
+                continue
             f = open(f"{self.output_path}/{query_names[i]}.txt", "w")
             f.write("Name     Distance" + "\n")
